Solve all problems. Left fix security filter.
