--- modules/core/src/com/haulmont/addon/cubajm/BeanPostProcessorImpl.java
@@ -1,36 +1,0 @@
-/*
- * Copyright (c) 2008-2021 Haulmont.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package com.haulmont.addon.cubajm;
-
-import com.haulmont.cuba.core.sys.singleapp.SingleAppCoreContextLoader;
-import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.config.BeanPostProcessor;
-import org.springframework.stereotype.Component;
-
-@Component
-public class BeanPostProcessorImpl implements BeanPostProcessor {
-    @Override
-    public Object postProcessBeforeInitialization(Object bean, String beanName) throws BeansException {
-        if (bean instanceof SingleAppCoreContextLoader) {
-            return new MySingleAppCoreServletLoader();
-        }
-        return bean;
-    }
-}
