--- modules/web/src/com/haulmont/addon/cubajm/DefaultWebFilterCreator.java
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2008-2021 Haulmont.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package com.haulmont.addon.cubajm;import net.bull.javamelody.MonitoringFilter;
-
-import javax.servlet.*;
-import javax.servlet.annotation.WebListener;
-import java.util.EnumSet;
-
-@WebListener
-public class DefaultWebFilterCreator implements ServletContextListener {
-    @Override
-    public void contextInitialized(ServletContextEvent sce) {
-        ServletContext servletContext = sce.getServletContext();
-        FilterRegistration.Dynamic javamelody = servletContext.addFilter("javamelody-web", MonitoringFilter.class);
-        javamelody.setInitParameter("monitoring-path", "/monitoring");
-        javamelody.setAsyncSupported(true);
-        javamelody.addMappingForUrlPatterns(
-                EnumSet.of(DispatcherType.REQUEST, DispatcherType.ASYNC), true, "/*");
-    }
-
-    @Override
-    public void contextDestroyed(ServletContextEvent sce) {}
-}
